KBC-2383 move assertMetadataEquals to MetadataUtils
